+++ pubspec.yaml
@@ -37,14 +37,19 @@ dependencies:
   flutter:
     sdk: flutter
 
+  injectable: ^2.1.1
+  get_it: ^7.2.0
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  injectable_generator: ^2.1.3
+  build_runner:
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
